Updated Hello_world.txt to have the UTF-8 BOM. This lets browsers properly display unicode.
Changed init.lua to a stub that launches the main code.

Added delay to the beginning init stage allow for easy recovery from misbehaving init script.

Moved main init code out to separate file, and run it from a stub. This allows for the meat of the init to be compiled if desired, greatly reducing size and increasing speed.

Added code to initialize GPIO to sane state from boot. Most useful for the garage_door_opener example.